layers/shared/nlb_common/requirements.txt
pyhumps
ujson